pyproject.toml
[tool.poetry]
name = "olympichistory"
version = "0.1.0"
description = ""
authors = ["Enias"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
dask = "^2021.9.1"
numpy = "^1.21.2"
pandas = "^1.3.3"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
